Refactor renderers and utility functions: standardize type usage, leverage `= default` for destructors, and improve return type safety

+++ wxChaos/core/SystemUtils.cpp
@@ -5,5 +5,5 @@ int Get_Cores()
 {
     SYSTEM_INFO sysinfo;
     GetSystemInfo(&sysinfo);
-    return sysinfo.dwNumberOfProcessors;
+    return static_cast<int>(sysinfo.dwNumberOfProcessors);
 }
